Created PlaceService interface and impl
